# backend/api-docs.yaml
openapi: 3.0.0
info:
  title: YouTube Clone API
  version: 1.0.0
  description: |
    Complete API documentation for YouTube Clone application.
    This API provides endpoints for user management, video uploads, and comments.
  contact:
    name: API Support
    email: support@youtubeclone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.youtubeclone.com
    description: Production server
...
paths:
...
  /api/v1/video/upload:
    post:
      tags:
        - Videos
      summary: Upload a new video
      description: Upload a video with thumbnail. Requires authentication. Videos are automatically approved.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - description
                - video
                - thumbnail
              properties:
                title:
                  type: string
                  example: "My Awesome Video"
                description:
                  type: string
                  example: "This is a great video"
                category:
                  type: string
                  example: "Entertainment"
                tags:
                  type: string
                  example: "funny, tutorial, tech"
                  description: Comma-separated tags
                video:
                  type: string
                  format: binary
                thumbnail:
                  type: string
                  format: binary
      responses:
        '201':
          description: Video uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Video uploaded successfully"
                  video:
                    $ref: '#/components/schemas/Video'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/all:
    get:
      tags:
        - Videos
      summary: Get all videos
      description: Retrieve all videos. No authentication required.
      responses:
        '200':
          description: List of all videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/my-videoes:
    get:
      tags:
        - Videos
      summary: Get my videos
      description: Get all videos uploaded by the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/update/{id}:
    put:
      tags:
        - Videos
      summary: Update video
      description: Update video metadata and/or thumbnail. Only the video owner can update.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Video ID
          example: "507f1f77bcf86cd799439011"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Video Title"
                description:
                  type: string
                  example: "Updated description"
                category:
                  type: string
                  example: "Education"
                tags:
                  type: string
                  example: "updated, tags"
                  description: Comma-separated tags
                thumbnail:
                  type: string
                  format: binary
                  description: Optional - new thumbnail image
      responses:
        '200':
          description: Video updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Video updated successfully"
                  video:
                    $ref: '#/components/schemas/Video'
        '403':
          description: Unauthorized - not the video owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/delete/{id}:
    delete:
      tags:
        - Videos
      summary: Delete video
      description: Delete a video. Only the video owner can delete.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Video ID
          example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Video deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Video deleted successfully"
        '403':
          description: Unauthorized - not the video owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/{id}:
    get:
      tags:
        - Videos
      summary: Get video by ID
      description: Get a specific video by ID. View count is incremented automatically. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Video ID
          example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Video retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/like:
    post:
      tags:
        - Videos
      summary: Like a video
      description: Like a video. If previously disliked, removes dislike and adds like.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - videoId
              properties:
                videoId:
                  type: string
                  example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Video liked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Video liked successfully"
                  video:
                    $ref: '#/components/schemas/Video'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/dislike:
    post:
      tags:
        - Videos
      summary: Dislike a video
      description: Dislike a video. If previously liked, removes like and adds dislike.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - videoId
              properties:
                videoId:
                  type: string
                  example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Video disliked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Video disliked successfully"
                  video:
                    $ref: '#/components/schemas/Video'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/tags/{tag}:
    get:
      tags:
        - Videos
      summary: Get videos by tag
      description: Retrieve all videos with a specific tag.
      parameters:
        - in: path
          name: tag
          required: true
          schema:
            type: string
          description: Tag name
          example: "tutorial"
      responses:
        '200':
          description: List of videos with the tag
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/category/{category}:
    get:
      tags:
        - Videos
      summary: Get videos by category
      description: Retrieve all videos in a specific category.
      parameters:
        - in: path
          name: category
          required: true
          schema:
            type: string
          description: Category name
          example: "Entertainment"
      responses:
        '200':
          description: List of videos in the category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/video/search:
    get:
      tags:
        - Videos
      summary: Search videos
      description: Search videos by title, description, tags, or category.
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Search query
          example: "tutorial"
        - in: query
          name: query
          required: false
          schema:
            type: string
          description: Alternative search query parameter
          example: "tutorial"
      responses:
        '200':
          description: List of matching videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
        '400':
          description: Search query is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== COMMENT ROUTES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        ChannelName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        logoUrl:
          type: string
          format: uri
        subscriber:
          type: number
          default: 0
        subscribedChannels:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
    Video:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        user_id:
          type: string
        category:
          type: string
          default: "General"
        tags:
          type: array
          items:
            type: string
        likes:
          type: array
          items:
            type: string
        dislikes:
          type: array
          items:
            type: string
        views:
          type: number
          default: 0
        status:
          type: string
          enum: [pending, approved, rejected]
          default: approved
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...